Premaknil alphabeta inteligenco v svoj razred
